tests: Added unit test for checklist use-cases

+++ api/src/application/use-cases/checklist/DeleteChecklistUseCase.js
@@ -4,7 +4,15 @@ class DeleteChecklistUseCase {
   }
 
   async execute(checklistId) {
-    return this.checklistRepository.delete(checklistId);
+    if (!checklistId) throw new Error('checklistId was not provided');
+
+    const result = await this.checklistRepository.delete(checklistId);
+
+    if (result === 0) {
+      throw new Error('Something went wrong removing the checklist');
+    }
+
+    return result;
   }
 }
